Defined Docker image for the API

+++ src/Presentation/Api/Core/Extensions/IServiceProvider.InitDb.cs
@@ -31,21 +31,26 @@ public static partial class Extensions {
     /// Initializes the database.
     /// </summary>
     /// <param name="self">The app builder.</param>
-    public static void InitDb(this WebApplication app) {
+    public static void InitDb(this WebApplication app, bool force = false) {
 
         using var scope = app.Services.CreateScope();
         app.Logger.LogInformation("Database initialization started...");
 
         var dbContext = scope.ServiceProvider.GetRequiredService<DbContext>();
-        dbContext.Database.EnsureDeleted();
-        dbContext.Database.EnsureCreated();
+
+        if (force) {
+
+            dbContext.Database.EnsureDeleted();
+            dbContext.Database.EnsureCreated();
         
-        dbContext.Add(Manager.TryCreate(app.Configuration["Root:Email"]!, app.Configuration["Root:Password"]!, ManagerPermissions.SuperAdminRole).Unwrap());
-        
-        dbContext.InitUsers();
-        dbContext.InitCategories();
-        dbContext.InitResources();
-        dbContext.InitComments();
+            dbContext.Add(Manager.TryCreate(app.Configuration["Root:Email"]!, app.Configuration["Root:Password"]!, ManagerPermissions.SuperAdminRole).Unwrap());
+            
+            dbContext.InitUsers();
+            dbContext.InitCategories();
+            dbContext.InitResources();
+            dbContext.InitComments();
+
+        } else dbContext.Database.EnsureCreated();
 
         app.Logger.LogInformation("Database initialization completed!");
